BigInteger in models id

+++ db/models.py
@@ -20,8 +20,8 @@ class User(DeclarativeBase):
 class Competitor(DeclarativeBase):
     __tablename__ = "competitors"
 
-    id = Column("competitor_id", BigInteger, nullable=False, primary_key=True)
-    user_id = Column(Integer, ForeignKey("users.user_id", ondelete="CASCADE"),
+    id = Column("competitor_id", Integer, nullable=False, primary_key=True)
+    user_id = Column(BigInteger, ForeignKey("users.user_id", ondelete="CASCADE"),
                      nullable=False)
     author_name = Column(String(100), nullable=False)
     client = Column(String(100), nullable=False)
